requirement.txt
tensorflow-gpu==1.15
scipy==1.7.3
Pillow==9.0.0
matplotlib==3.5.1
opencv-python==4.5.5.62
